requirements.txt
Django==3.1.4
django-import-export==2.4.0
